refact: move response file to appropriate directories
