+++ src/main/kotlin/com/dobby/backend/domain/gateway/S3Gateway.kt
@@ -1,6 +1,6 @@
 package com.dobby.backend.domain.gateway
 
-import com.dobby.backend.domain.dto.PreSignedUrlResponse
+import com.dobby.backend.presentation.api.dto.response.PreSignedUrlResponse
 
 interface S3Gateway {
     fun getPreSignedUrl(fileName: String): PreSignedUrlResponse
