+++ src/main/kotlin/com/dobby/backend/presentation/api/dto/response/PreSignedUrlResponse.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.domain.dto
+package com.dobby.backend.presentation.api.dto.response
 
 data class PreSignedUrlResponse(
     val url: String
